Change humanNumber prototype to take int64. It doesn't actually do anything useful with them yet, but we'll see about that.
p4raw-id: //oryx/src/main@53075

+++ core/string.cpp
@@ -1480,7 +1480,7 @@ void String::print() const
     The number is rounded more or less correctly.
 */
 
-String String::humanNumber( uint n )
+String String::humanNumber( int64 n )
 {
     if ( n < 1024 )
         return fromNumber( n );
